Move ActiveStorage reflection logic entirely into ActiveStorage
